+++ src/mem/Makefile.mk
@@ -5,5 +5,7 @@
 
 abt_sources += \
 	mem/malloc.c \
+	mem/malloc_task.c \
+	mem/malloc_thread.c \
 	mem/valgrind.c
